TEST: Enable some tests for simple-olog
