+++ packages/logics/tsconfig.test.json
@@ -5,5 +5,5 @@
         "types": ["node"],
         "lib": ["ES2023", "DOM", "DOM.Iterable"]
     },
-    "include": ["vite.config.ts", "test/**/*.ts"]
+    "include": ["vite.config.ts", "test/simple-olog.test.ts"]
 }
